added components and fixed filtration
