+++ src/App.tsx
@@ -1,10 +1,13 @@
 /* eslint-disable jsx-a11y/label-has-associated-control */
 /* eslint-disable jsx-a11y/control-has-associated-label */
-import React, { useState, useEffect, useRef } from 'react';
-import classNames from 'classnames';
+import React, { useState, useEffect } from 'react';
 
 import { UserWarning } from './UserWarning';
 import { USER_ID } from './api/todos';
+import { Header } from './components/Header/Header';
+import { TodoList } from './components/TodoList/TodoList';
+import { Footer } from './components/Footer/Fotter';
+import { Error } from './components/Error/Erros';
 import { Todo } from './types/Todo';
 import { getTodos } from './api/todos';
 
@@ -23,38 +26,19 @@ function filterTodos(todos: Todo[], status?: string | null) {
 }
 
 export const App: React.FC = () => {
-  const [query, setQuery] = useState('');
   const [todos, setTodos] = useState<Todo[]>([]);
-  const [filteredTodos, setFilteredTodos] = useState<Todo[]>([]);
   const [errorMessage, setErrorMessage] = useState('');
-  const currentList = useRef('all');
-  const areAllCompleted = todos.every(todo => todo.completed);
-  const isAnyCompleted = todos.some(todo => todo.completed);
+  const [status, setStatus] = useState('all');
   const noTodos = todos.length === 0;
-  const activeTodos = todos.filter(todo => !todo.completed);
+  const filteredTodos = filterTodos(todos, status);
 
   useEffect(() => {
     getTodos()
       .then(response => {
         setTodos(response);
-        setFilteredTodos(response);
       })
       .catch(() => setErrorMessage('Unable to load todos'));
-
-    setTimeout(() => setErrorMessage(''), 3000);
   }, []);
-
-  const handleFiltration = (
-    e: React.MouseEvent<HTMLAnchorElement, MouseEvent>,
-  ) => {
-    const status = e.currentTarget.textContent?.toLowerCase();
-
-    setFilteredTodos(filterTodos(todos, status));
-
-    if (status) {
-      currentList.current = status;
-    }
-  };
 
   if (!USER_ID) {
     return <UserWarning />;
@@ -65,155 +49,16 @@ export const App: React.FC = () => {
       <h1 className="todoapp__title">todos</h1>
 
       <div className="todoapp__content">
-        <header className="todoapp__header">
-          {!noTodos && (
-            <button
-              type="button"
-              className={classNames('todoapp__toggle-all', {
-                active: areAllCompleted,
-              })}
-              data-cy="ToggleAllButton"
-            />
-          )}
+        <Header todos={todos} />
 
-          {/* Add a todo on form submit */}
-          <form>
-            <input
-              data-cy="NewTodoField"
-              type="text"
-              className="todoapp__new-todo"
-              placeholder="What needs to be done?"
-              value={query}
-              onChange={event => setQuery(event.target.value)}
-              autoFocus
-            />
-          </form>
-        </header>
-
-        <section className="todoapp__main" data-cy="TodoList">
-          {filteredTodos.map(todo => {
-            return (
-              <div
-                data-cy="Todo"
-                className={classNames('todo', {
-                  completed: todo.completed,
-                  'todo__status-label': !todo.completed,
-                })}
-                key={todo.id}
-              >
-                <label className="todo__status-label">
-                  <input
-                    data-cy="TodoStatus"
-                    type="checkbox"
-                    className="todo__status"
-                    checked={todo.completed}
-                  />
-                </label>
-
-                <span data-cy="TodoTitle" className="todo__title">
-                  {todo.title}
-                </span>
-
-                {/* Remove button appears only on hover */}
-                <button
-                  type="button"
-                  className="todo__remove"
-                  data-cy="TodoDelete"
-                >
-                  ×
-                </button>
-
-                {/* overlay will cover the todo while it is being deleted or updated */}
-                <div
-                  data-cy="TodoLoader"
-                  className={classNames('modal', 'overlay')}
-                >
-                  <div className="modal-background has-background-white-ter" />
-                  <div className="loader" />
-                </div>
-              </div>
-            );
-          })}
-        </section>
-
-        {/* Hide the footer if there are no todos */}
+        <TodoList filteredTodos={filteredTodos} />
 
         {!noTodos && (
-          <footer className="todoapp__footer" data-cy="Footer">
-            <span className="todo-count" data-cy="TodosCounter">
-              {`${activeTodos.length} items left`}
-            </span>
-
-            {/* Active link should have the 'selected' class */}
-            <nav className="filter" data-cy="Filter">
-              <a
-                href="#/"
-                className={classNames('filter__link', {
-                  selected: currentList.current === 'all',
-                })}
-                data-cy="FilterLinkAll"
-                onClick={event => handleFiltration(event)}
-              >
-                All
-              </a>
-
-              <a
-                href="#/active"
-                className={classNames('filter__link', {
-                  selected: currentList.current === 'active',
-                })}
-                data-cy="FilterLinkActive"
-                onClick={event => handleFiltration(event)}
-              >
-                Active
-              </a>
-
-              <a
-                href="#/completed"
-                className={classNames('filter__link', {
-                  selected: currentList.current === 'completed',
-                })}
-                data-cy="FilterLinkCompleted"
-                onClick={event => handleFiltration(event)}
-              >
-                Completed
-              </a>
-            </nav>
-
-            {/* this button should be disabled if there are no completed todos */}
-            <button
-              type="button"
-              className="todoapp__clear-completed"
-              data-cy="ClearCompletedButton"
-              disabled={!isAnyCompleted}
-            >
-              Clear completed
-            </button>
-          </footer>
+          <Footer todos={todos} status={status} onStatusChange={setStatus} />
         )}
       </div>
 
-      {/* DON'T use conditional rendering to hide the notification */}
-      {/* Add the 'hidden' class to hide the message smoothly */}
-      <div
-        data-cy="ErrorNotification"
-        className={classNames(
-          'notification',
-          'is-danger',
-          'is-light',
-          'has-text-weight-normal',
-          { hidden: !errorMessage },
-        )}
-      >
-        <button
-          data-cy="HideErrorButton"
-          type="button"
-          className="delete"
-          onClick={() => setErrorMessage('')}
-        />
-        {/* show only one message at a time */}
-        {errorMessage}
-      </div>
+      <Error errorMessage={errorMessage} hideError={setErrorMessage}/>
     </div>
   );
 };
