+++ src/components/Footer/Fotter.tsx
@@ -1,0 +1,69 @@
+import React from 'react';
+import classNames from 'classnames';
+
+import { Todo } from '../../types/Todo';
+
+type Props = {
+  todos: Todo[];
+  status: string;
+  onStatusChange: (arg: string) => void;
+};
+
+export const Footer: React.FC<Props> = ({ todos, status, onStatusChange }) => {
+  const activeTodos = todos.filter(todo => !todo.completed);
+  const isAnyCompleted = todos.some(todo => todo.completed);
+
+  return (
+    <footer className="todoapp__footer" data-cy="Footer">
+      <span className="todo-count" data-cy="TodosCounter">
+        {`${activeTodos.length} items left`}
+      </span>
+
+      {/* Active link should have the 'selected' class */}
+      <nav className="filter" data-cy="Filter">
+        <a
+          href="#/"
+          className={classNames('filter__link', {
+            selected: status === 'all',
+          })}
+          data-cy="FilterLinkAll"
+          onClick={() => onStatusChange('all')}
+        >
+          All
+        </a>
+
+        <a
+          href="#/active"
+          className={classNames('filter__link', {
+            selected: status === 'active',
+          })}
+          data-cy="FilterLinkActive"
+          onClick={() => onStatusChange('active')}
+        >
+          Active
+        </a>
+
+        <a
+          href="#/completed"
+          className={classNames('filter__link', {
+            selected: status === 'completed',
+          })}
+          data-cy="FilterLinkCompleted"
+          onClick={() => onStatusChange('completed')}
+        >
+          Completed
+        </a>
+      </nav>
+
+      {/* this button should be disabled if there are no completed todos */}
+      <button
+        type="button"
+        className="todoapp__clear-completed"
+        data-cy="ClearCompletedButton"
+        disabled={!isAnyCompleted}
+      >
+        Clear completed
+      </button>
+    </footer>
+  );
+};
