+++ src/components/Header/Header.tsx
@@ -1,0 +1,42 @@
+import React from 'react';
+import { useState } from 'react';
+import classNames from 'classnames';
+
+import { Todo } from '../../types/Todo';
+
+type Props = {
+  todos: Todo[];
+};
+
+export const Header: React.FC<Props> = ({ todos }) => {
+  const [query, setQuery] = useState('');
+  const areAllCompleted = todos.every(todo => todo.completed);
+  const noTodos = todos.length === 0;
+
+  return (
+    <header className="todoapp__header">
+      {!noTodos && (
+        <button
+          type="button"
+          className={classNames('todoapp__toggle-all', {
+            active: areAllCompleted,
+          })}
+          data-cy="ToggleAllButton"
+        />
+      )}
+
+      {/* Add a todo on form submit */}
+      <form>
+        <input
+          data-cy="NewTodoField"
+          type="text"
+          className="todoapp__new-todo"
+          placeholder="What needs to be done?"
+          value={query}
+          onChange={event => setQuery(event.target.value)}
+          autoFocus
+        />
+      </form>
+    </header>
+  );
+};
