+++ src/components/TodoList/TodoList.tsx
@@ -1,0 +1,54 @@
+import React from 'react';
+import classNames from 'classnames';
+
+import { Todo } from '../../types/Todo';
+
+type Props = {
+  filteredTodos: Todo[];
+};
+
+export const TodoList: React.FC<Props> = ({ filteredTodos }) => {
+  return (
+    <section className="todoapp__main" data-cy="TodoList">
+      {filteredTodos.map(todo => {
+        return (
+          <div
+            data-cy="Todo"
+            className={classNames('todo', {
+              completed: todo.completed,
+              'todo__status-label': !todo.completed,
+            })}
+            key={todo.id}
+          >
+            <label className="todo__status-label">
+              <input
+                data-cy="TodoStatus"
+                type="checkbox"
+                className="todo__status"
+                checked={todo.completed}
+              />
+            </label>
+
+            <span data-cy="TodoTitle" className="todo__title">
+              {todo.title}
+            </span>
+
+            {/* Remove button appears only on hover */}
+            <button type="button" className="todo__remove" data-cy="TodoDelete">
+              ×
+            </button>
+
+            {/* overlay will cover the todo while it is being deleted or updated */}
+            <div
+              data-cy="TodoLoader"
+              className={classNames('modal', 'overlay')}
+            >
+              <div className="modal-background has-background-white-ter" />
+              <div className="loader" />
+            </div>
+          </div>
+        );
+      })}
+    </section>
+  );
+};
